requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
pytest==8.3.5
python-dotenv==1.0.1
waitress==3.0.2
